Factor out singletonformset_factory(), SingletonFormsetFormMutation.
